Bug Fixing, Made they keycard puzzle easier
Gave the walls in the last room bigger hitboxes to prevent a bug where the boxes are pushed through the walls.
Gave the keycard a bigger collider so the player is able to pick it up when standing close to the seat, rather than on top of it after receiving feedback that it was too hard to figure out.